# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  domain.Save:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      last_modified:
        type: string
      name:
        type: string
      size:
        type: integer
      updated_at:
        type: string
    type: object
...
  errors.AppError:
    properties:
      code:
        description: 错误码
        type: integer
      data:
        description: 返回数据
      message:
        description: 错误信息
        type: string
    type: object
  handler.ApplySaveRequest:
    properties:
      save_id:
        type: integer
      server_id:
        type: integer
    required:
    - save_id
    - server_id
    type: object
...
host: localhost:11223
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: MC Dashboard 后端服务API文档
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: MC Dashboard API
  version: "1.0"
paths:
...
  /saves:
    get:
      consumes:
      - application/json
      description: 获取所有存档文件列表
      parameters:
      - default: 0
        description: 页码，从1开始
        in: query
        name: page
        type: integer
      - default: 0
        description: 每页数量
        in: query
        name: size
        type: integer
      - default: desc
        description: 排序方向，asc或desc
        in: query
        name: order
        type: string
      - default: create_at
        description: 排序字段，支持id/name/version
        in: query
        name: orderBy
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.Save'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.AppError'
      security:
      - Bearer: []
      summary: 获取存档列表
      tags:
      - 存档管理
    post:
      consumes:
      - multipart/form-data
      description: 上传新的存档文件
      parameters:
      - description: 存档文件
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Save'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.AppError'
      security:
      - Bearer: []
      summary: 上传存档
      tags:
      - 存档管理
  /saves/{id}:
    delete:
      consumes:
      - application/json
      description: 删除指定的存档文件
      parameters:
      - description: 存档id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/errors.AppError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.AppError'
      security:
      - Bearer: []
      summary: 删除存档
      tags:
      - 存档管理
  /saves/apply:
    post:
      consumes:
      - application/json
      description: 将指定的存档应用到服务器
      parameters:
      - description: 存档应用参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.ApplySaveRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/errors.AppError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.AppError'
      security:
      - Bearer: []
      summary: 应用存档
      tags:
      - 存档管理
...
schemes:
- http
- https
securityDefinitions:
  Bearer:
    description: 请输入Bearer token，格式为：Bearer <token>
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"